+++ tests/Agent/IntegrationTests/UnboundedServices/mssql/restore.sql
@@ -2,7 +2,7 @@ USE [master]
 GO
 RESTORE DATABASE [NewRelic]
 FROM DISK = '/var/opt/mssql/backup/NewRelicDB.bak'
-WITH
+WITH REPLACE,
 MOVE 'NewRelic' TO '/tmp/NewRelic.mdf',
 MOVE 'NewRelic_log' TO '/tmp/NewRelic_log.ldf';
 GO