+++ tests/Agent/PerformanceTests/PerformanceTestApp/Program.cs
@@ -11,6 +11,10 @@ var builder = WebApplication.CreateBuilder(args);
 
 builder.Host.UseSerilog((_, config) => config.WriteTo.Console());
 
+// Print .NET version
+var dotnetVersion = Environment.Version.ToString();
+Console.WriteLine("Starting PerformanceTestApp on .NET {0}", dotnetVersion);
+
 builder.Services.AddControllers();
 
 var mongoConnectionString = builder.Configuration["MONGODB_CONNECTION_STRING"]
@@ -25,7 +29,7 @@ builder.Services.AddSingleton<IConnectionMultiplexer>(
 var rabbitHost = builder.Configuration["RABBITMQ_HOST"] ?? "localhost";
 var rabbitConnection = new ConnectionFactory { HostName = rabbitHost }.CreateConnection();
 using (var ch = rabbitConnection.CreateModel())
-    ch.QueueDeclare("perf", durable: false, exclusive: false, autoDelete: false, arguments: null);
+    ch.QueueDeclare("perf", durable: true, exclusive: false, autoDelete: false, arguments: null);
 builder.Services.AddSingleton<IConnection>(rabbitConnection);
 
 builder.WebHost.UseUrls($"http://{IPAddress.Any}:8080");